Correct links to advanced filters in docstrings?
This is a possible correction to #124: `mkdocs` complains about not
recognizing the relative links' destination. The recommendation it
provides didn't quite make sense to me relative to the URL structure as
rendered, but it apparently automatically substituted it in and it
appeared to work, so I just left them as I'd written them.

Now, I realize now that it may be more about the actual structure of the
filesystem getting from `api/fastcrud.md` (which contains the includes
from the code's docstrings) to `advanced/crud.md`, and `mkdocs` handles
translating the file paths to URLs. So I put this PR together to fix the
links and you can figure out whether or not it needs to go in. :D